third_party/blink/web_tests/fast/html/imports/import-dynamic-tree-expected.txt

CONSOLE WARNING: line 14: HTML Imports is deprecated and will be removed in M80, around February 2020. Please use ES modules instead. See https://www.chromestatus.com/features/5144752345317376 and https://developers.google.com/web/updates/2019/07/web-components-time-to-upgrade for more details.
This test ensures that script-added import lets its child blocks following scripts.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS window.trace is ['START root', 'child1', 'child2', 'END root']
PASS successfullyParsed is true

TEST COMPLETE
